requirements.txt
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.6
executing==1.2.0
Flask==2.2.3
fonttools==4.39.3
gensim==4.3.1
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
language-tool-python==2.7.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
packaging==23.1
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pyenchant==3.2.2
Pygments==2.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
WTForms==3.0.1